webhooks: Deterministically match webhook events to repos
Previously, we would attempt to recreate the clone URL of a repo based on event data. This is a lossy matching, and can cause events to get rejected, although we have the repo cloned.

This PR changes the matching to instead use the external ID of the repo, which we already store in the repo table in a separate column.

Closes SRC-40

Test plan:

Tests still pass, set up webhooks locally and they still matched (but only tried GitHub).