fix(svelte): Render correct error page (#63663)
Details about the probblem can be found in srch-527 (fixes srch-527).

This commit implements a workaround via the `handleError` hook. If this
function returns an error then SvelteKit is using that instead of its
own error object. So we unwrap the error object ourselves (since
SvelteKit fails to do that due to bazel build issues) and ensure this
way that the correct error is passed to our error template.

## Test plan

Coincidentally we disabled all error related tests in CI, so by enabling
these tests again we should see whether this approach works.