fix(svelte): Fixes excessive 'the hotkey ... has already been registered' messages (#63679)
Fixes srch-697

This commit makes two important changes:

- The popover component will now only register the ESC key handler when
it is open (und unregister when it closes). This was done by exposing a
new function that creates a hotkey without registering it. The existing
`registerHotkey` function simply wraps the new function so that existing
callsites are not affected by this change.
- The error message was turned into a warning and will now only be
triggered in dev builds.

## Test plan

Existing unit test, manual testing.