feat(svelte): Add cody button to repo home page
This adds the cody button to the repe "home page" so that cody can be
opened from there.
This commit also adds a fixed height to the various headers so that they
align well with cody (and don't change size when the cody button is
hidden when the sidebar is open).