Search: restore snippets to filters sidebar (#63587)
This adds snippets back to the search sidebar, which got missed when it
was redesigned.

This includes some refactoring of the Svelte version to account for
filter types that do not match 1:1 with the backend types. We initially
tried to tie them tightly with the backend types so the backend is the
source of truth, but I think we want to have the ability to introduce
client-side-only filters, which we already sorta hackily do with the
`type:` filters. And it's even more hacky with the `count:` filter in
the React webapp (which doesn't look like it was ever implemented in the
Svelte version).

In the React app, I made the minimum changes to get this working (no
associated refactoring).