chore(appliance): extract envtest helper package (#63689)
**chore(appliance): remove dependency of envtest helpers on testify**

In preparation for extracting them for reuse.



**chore(appliance): extract envtest helper package**

In preparation for reuse.

The motivation for this refactor is to be able to write tests against
objects pushed to a real kubernetes API in packages other than
appliance/reconciler. For example, so that we can assert on the shape of
the ConfigMap created by the appliance API handlers.