Cody Web: bump cody web to 0.2.3 (#63723)
Updating the `cody-web-experimental` package to the latest version which
- Fixed problems with telemetry (now we send agent level of telemetry
events with `Web.Cody` client name)
- Fixed remote repository context as you switch between chats 
- Added support for cody context filters in file and symbols mentions 
- Improved link rendering for remote files (no more "remote-file://"
protocol prefixes.

Also, this PR adds some missing CSS token overrides for the LLM picker
on dotcom.

Note: I, by accident, pushed the `0.2.2` version with the same build as
I have published already in `0.2.1`. NPM doesn't allow to override the
build with the same version, so I had to publish 0.2.3 to make it
through to the Sourcegraph repo. This is why we jump from 0.2.1 to 0.2.3

## Test plan
-General manual checks for Cody Web