feat(svelte): Add reblame support
This commit adds a link next to the blame commit message that allows
reblaming to a prior commit if available.

It extends the existing blame extension. This won't have any affect on
the React app because it doesn't pass the configuration option needed to
add the extra gutter.

Some notes:

- I originally used the tooltip component instead of `title` but somehow
  it starts to break when scrolling the document (tooltips don't show up
  anymore). I don't know if CodeMirror does anything to the DOM elements
  that causes this to fail.
- The reblame URL also selected the corresponding line so that the
  correct line is scrolled into view.