fix(svelte): Add back default context:global filter
Currently we don't show the global context filter on the search home
page or the search results page (global context is the default context).

This commit does two things:

- It prefills the search input on the search homepage with the user's
  default context (like the React app)
- It the logic that pre-processed the search query and removed the
  context filter if it was global.

In other words we simplify the query logic by showing/submitting the
search query as is. Noteably this doesn't affect how the search input
works on repo pages.