fix/internal/txemail: add timeout for SMTP connection establishment (#63759)
Before, the smtp.Dial function had no cancellation mechanism - meaning
that you could be waiting for several minutes to try to establish a
connection before it gives up.

I work around this by establishing a TCP connection myself to the
appropriate address, and using
[net.DialContext](https://pkg.go.dev/net#Dialer.DialContext) to cancel
the dial process if either:

1. The parent context cancels
2. 30 seconds have passed 

Afterwards, we construct the smtp.Client ourselves using
[smtp.NewClient](https://pkg.go.dev/net/smtp#NewClient)

This only changes what happens when we try to establish a connection -
it doesn't change any behavior around sending emails afterwards.

## Test plan

Manual testing. 

1. On local dev, I edited the credentials in the [dev-private site
configuration file
](https://github.com/sourcegraph/dev-private/blob/67ee9b3a69d6dcfc6b4558fe780d301da70774ee/enterprise/dev/site-config.json#L8-L11)to
point to an invalid port.

2. I attempted to create an access token, and after only30 seconds I saw
this error print in the terminal:

```
[       frontend] WARN schemaResolver.CreateAccessToken graphqlbackend/access_tokens.go:127 Failed to send email to inform user of access token creation {"userID": 1, "error": "establishing TCP connection to \"smtp.gmail.com:57\": dial tcp 74.125.137.109:57: i/o timeout"}
```

## Changelog

- Instead of waiting forever, we wait at most 30 seconds before giving
up when trying to connect to the configured mail server when sending an
email.