feat/graphql/access_tokens: send access token creation/deletion emails in background and create special message for dial errors (#63760)
Closes
https://linear.app/sourcegraph/issue/SRC-463/error-handling-for-smtp-config-issues

Before, when someone created or deleted an access token, we'd block on
sending the email before we'd return to the user.

However, if an error occurred when sending an email - the error was only
logged - **not** returned to the caller.

This PR makes three changes:

1. We now sending the email in a background goroutine. This makes it so
that the caller doesn't have to wait for the email to be sent before
getting the access token back. This isn't a behavior change since we
only logged any SMTP errors and **deliberatly avoided** forwarding that
error to the caller. That behavior doesn't change here.

1. If an error occurs while sending the access token email, we now have
a specialized log message indicating that the user's SMTP configuration
might be incorrect.

1. The log error that's printed is now an `ERROR` instead of a `WARN`
level.

## Test plan

1. On local dev, I edited the credentials in the [dev-private site
configuration file
](https://github.com/sourcegraph/dev-private/blob/67ee9b3a69d6dcfc6b4558fe780d301da70774ee/enterprise/dev/site-config.json#L8-L11)to
point to an invalid port.

2. I attempted to create an access token, and after only 30 seconds I
saw this error print in the terminal:

```
[       frontend] WARN schemaResolver.CreateAccessToken graphqlbackend/access_tokens.go:150 Failed to send email to inform user of access token creation. (This error might indicate that your SMTP connection settings are incorrect. Please check your site configuration.) {"userID": 1, "error": "establishing TCP connection to \"smtp.gmail.com:57\": dial tcp 142.250.101.109:57: i/o timeout"}
```

3. I fixed the broken SMTP configuration, and deleted the access token.
I saw it suscessfully send an email:

![Screenshot 2024-07-10 at
11.44.07 AM.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/5VKJ5spRdhDRvKQ0TTIe/406e4b23-121f-4d47-a818-430e4ce1b656.png)


## Changelog

- When creating or deleting an access token, we no longer wait for the
email to be sent before returning to the caller. Instead, we now send it
in the background.