feat(svelte): Add search query syntax introduction (#63753)
Closes srch-494

This adds the search query syntax introduction component to the search
homepage. I tried to replicate the React version as closely as possible.

I originally wanted to reuse the logic to generate the example sections
but since it had dependencies on wildcard I duplicated it instead.

Notable additional changes:
- Added a `value` method to the temporary settings store to make it
easier to get the current value of the settings store. It only resolves
(or rejects) once the data is loaded.
- Extended the tabs component to not show the tab header if there is
only a single panel. This makes it easier for consumers to render tabs
conditionally.
- Added the `ProductStatusBadge` component
- Various style adjustments

For reference, the relevant parts of the React version are in
https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/client/branded/src/search-ui/components/useQueryExamples.tsx
and
https://sourcegraph.sourcegraph.com/github.com/sourcegraph/sourcegraph/-/blob/client/branded/src/search-ui/components/QueryExamples.tsx

## Test plan

Manual testing. I manually set the value received from temporary
settings to `null` (in code) to force trigger the compute logic.