fix(config): homogeneize desc for REDIS_ENDPOINT
PR 63625 introduced an explicit REDIS_ENDPOINT setting for the Cody
Gateway, whose description differ from the default in redispool.

internal/env.Get detects that and panics, which this commit fixes.

Ideally, we shouldn't read env var from internal packages, will submit a
ticket for that (and it should be caught by CI as well, when descs are
drifting).