remove ?visible=N param from some connection pages
This was intended to make it so that if you reloaded or navigated directly to a page that had a list, subsequently fetching the next page would only get `pageSize` additional records (eg 20), not `first` (which is however many records were already showing) for batch-based navigation. This is not worth the additional complexity that `?visible=` introduces, and is not clearly desirable even.