chore(svelte): Refactor repo loaders (#63828)
This commit is in preparation for other work. It refactors how we work
with resolved repository and resolved revision information.

The idea was that the top level repo loader should try to resolve the
repository information and error if that wasn't possible. Not being able
to resolve the revision was accepted, hence this check:

```js
// still render the main repo navigation and header
if (!isRevisionNotFoundErrorLike(repoError)) {
  error(400, asError(repoError))
}
```

However the way it was implemented meant that we wouldn't pass any
resolved repository information to the sub-pages/layouts when the
revision couldn't be resolved, which seems wrong.

With these changes, the top level repo loader now provides a
`resolvedRepository` object (where `commit`/`changelist` might be unset)
and the `(validrev)` loader creates the `resolvedRevision` object, just
how its sub-pages/layouts expect.

And instead of returning error objects from `resolveRepoRevision` and
checking them in the loader we throw errors/redirects directly in that
function. IMO that makes the whole flow easier to understand.

## Test plan

Manual testing, CI integration tests